refactor(secrets): split the module along the up and down paths
`secrets/mod.rs` reached 591 code lines against the 500 hard limit, and
the reason it grew is that it had been holding three separate jobs:
assembling per-service references, creating the project's secrets at
`up`, and tearing them down at `down`.

Split along those seams rather than by code-versus-tests, so each module
carries its own tests the way `plan.rs` already did:

  mod.rs     143  module docs, build_native_secrets, collect_payload_union
  create.rs  238  create_project_secrets, create_secret
  remove.rs  243  remove_internal_secrets and the teardown helpers

No behaviour change: the same 2007 tests pass, and the two entry points
keep the visibility they had — `pub(super)` meant `crate::engine` from
mod.rs and means `secrets` from a submodule, so both are now spelled
`pub(in crate::engine)` explicitly.

The two test helpers the create and remove suites share moved to a
`tests_support` module rather than being duplicated.
